Added a confirm msg to exercise 2.14 code

+++ phonebook/src/App.jsx
@@ -35,7 +35,10 @@ function App() {
   };
 
   const deletePerson = id => {
-    peopleService.deletePerson(id);
+    const userToDelete = persons.find(person => person.id == id)
+    if (confirm(`Are you sure you want to delete ${userToDelete.name}`)){
+      peopleService.deletePerson(id);
+    }
   }
 
   return (
